+++ src/main/java/com/javasleuth/security/InputValidator.java
@@ -28,7 +28,7 @@ public class InputValidator {
 
     // Commands that require special permission validation
     private static final Set<String> PRIVILEGED_COMMANDS = new HashSet<>(Arrays.asList(
-        "redefine", "retransform", "heapdump", "mc", "reset", "stop"
+        "redefine", "retransform", "heapdump", "mc", "reset", "stop", "vmtool"
     ));
 
     private final ProductionConfig config;